// src/app/model/ICandidateModel.ts
export interface ICandidateModel {
    firstName: string;
    lastName: string;
    email: string;
    mobileNumber: string;
    identityNumber: string;
    addressLine1: string;
    addressLine2: string;
    country: string;
    city: string;
    province: string;
    postalCode: string;
    linkedinProfile: string;
    facebookProfile: string;
    twitterProfile: string;
}